requirements.txt
aiofiles==23.1.0
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
boto3==1.26.122
botocore==1.29.122
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
debugpy==1.5.1
decorator==5.1.1
einops==0.6.1
encodec==0.1.1
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.4.0
funcy==2.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
install==1.3.5
ipykernel==6.15.0
ipython==8.12.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pip==23.0.1
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
pywin32==305.1
PyYAML==6.0
pyzmq==23.2.0
regex==2023.3.23
requests==2.29.0
s3transfer==0.6.0
scipy==1.10.1
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
stack-data==0.6.2
starlette==0.26.1
suno-bark==0.0.1a0
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
wheel==0.38.4
widgetsnbextension==4.0.7
zipp==3.15.0
fastapi
uvicorn
soundfile
python-multipart